% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Month_difference_calculation.R
\name{monthDiff}
\alias{monthDiff}
\title{Month difference calculation between two dates}
\usage{
monthDiff(startDate, endDate)
}
\arguments{
\item{startDate}{A first date as \%d/\%m/\%Y format}

\item{endDate}{A last date as \%d/\%m/\%Y format}
}
\value{
A number of month difference between a starting date and a final date
}
\description{
Month difference calculation between two dates
}
\examples{
monthDiff(d1, d2)
}
\author{
J.Chanthercrob
}
