% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Year_difference_calculation.R
\name{yearDiff}
\alias{yearDiff}
\title{Year difference calculation between two dates}
\usage{
yearDiff(startDate, endDate)
}
\arguments{
\item{startDate}{A first date as \%d/\%m/\%Y format}

\item{endDate}{A last date as \%d/\%m/\%Y format}
}
\value{
A number of year difference between a starting date and a final date
}
\description{
Year difference calculation between two dates
}
\examples{
yearDiff(d1, d2)
}
\author{
J.Chanthercrob
}
